refactor: refactoring files for tanstack
